pyproject.toml
[tool.poetry]
name = "cits3403-project"
version = "0.1.0"
description = ""
authors = [
    "Laser Racer <cat@heaven.ist>",
    "Neutronito <vg.sinalco@gmail.com>",
    "abde2 one <archie.tiz@gmail.com>"
]
packages = [
    { include = "flask_app", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.3"
python-dotenv = "^0.19.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Login = "^0.6.0"
Flask-WTF = "^1.0.1"
Flask-Scrypt = "^0.1.3"


[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-flask = "^1.2.0"
requests = "^2.27.1"

[tool.poetry.scripts]
flask-app = "flask_app:main"
flask-shell = "flask_app:shell"
flask-create-admin = "flask_app:create_admin"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
